[core] Move SymbolInstance and SymbolFeature into their own files
